Scratch buffer initialization in `run_aero_with_deform`
This patch changes how some local scratch buffers are initialized so
that the input type is taken into account to support e.g. dual numbers.
Initialization is now done by calling `zeros` directly which save some
allocations (the old code allocated once for the call to getindex
([...]), once in `copy` and finally once in the multiplication with 0).
Extracted from #100.